texel import loop fixes

+++ engine/evaluate.c
@@ -27,7 +27,7 @@ int i;
 return i;
 }
 
-int eval_phase(board *b){
+int8_t eval_phase(board *b){
 int i, tot, faze, q;
 int vaha[]={0,1,1,2,2,4};
 int nc[]={16,4,4,4,2};
